backend/app/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
alembic
python-dotenv
google-generativeai
supabase
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
spotipy
requests
